It works! (#537)
