+++ deployer/Cargo.toml
@@ -11,6 +11,8 @@ repository = "https://github.com/commonwarexyz/monorepo/tree/main/deployer"
 documentation = "https://docs.rs/commonware-deployer"
 
 [dependencies]
+commonware-cryptography = { workspace = true }
+commonware-utils = { workspace = true }
 aws-config = "1.5.16"
 aws-sdk-ec2 = "1.110.0"
 clap = "2.34"
@@ -18,7 +20,6 @@ futures = "0.3"
 reqwest = { version = "0.11" }
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"
-tempdir = "0.3"
 tokio = { version = "1.0", features = ["full"] }
 uuid = { version = "1.0", features = ["v4"] }
